adjusted case lammps
Former-commit-id: d4a0fa2a78eb076ef05945988a5d1426693bed25

+++ tutorials/walkthrough_notebooks/Liquid_Argon/from_traj/conductivity_gk.py
@@ -4,13 +4,13 @@ import shutil
 
 import mdsuite as mds # Cool name for professionalism purposes
 
-new_case = True
+new_case = False
 
-try:
-    new_case = True
-    shutil.rmtree('Argon')
-except FileNotFoundError:
-    pass
+if new_case:
+    try:
+        shutil.rmtree('Argon')
+    except FileNotFoundError:
+        pass
 
 
 
@@ -21,7 +21,7 @@ if new_case:
 
 # argon.help_computations_args('EinsteinDiffusionCoefficients') # auxiliary function to help on the extra arguments
 # argon.run_computation('EinsteinDiffusionCoefficients', species=list(argon.species.keys()),data_range=50, plot=False, singular=True, distinct=False)
-argon.run_computation('GreenKuboThermalConductivity', data_range=8000, plot=False)
+argon.run_computation('GreenKuboThermalConductivity', data_range=12000, plot=True)
 # argon.help_computations_args('GreenKuboDiffusionCoefficients')
 print(argon.results)
-argon.dump_results_json()
+# argon.dump_results_json()
